refactor(dashboards-list-v2): extract lastUpdatedLabel into utils
Dedupe the relative-time formatter that was copied into both DashboardRow
and ConfigureMetadataModal. Single source in DashboardsListPageV2/utils.ts.
